Adjusted exceptions models and handler
