+++ src/main/java/br/com/wferreiracosta/louis/controllers/UserMerchantsController.java
@@ -1,8 +1,8 @@
 package br.com.wferreiracosta.louis.controllers;
 
+import br.com.wferreiracosta.louis.exceptions.ValidationError;
 import br.com.wferreiracosta.louis.models.dtos.UserDTO;
 import br.com.wferreiracosta.louis.models.entities.UserEntity;
-import br.com.wferreiracosta.louis.models.exceptions.ValidationError;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.media.Content;
 import io.swagger.v3.oas.annotations.media.Schema;
