+++ src/main/java/br/com/wferreiracosta/louis/controllers/WalletController.java
@@ -1,7 +1,7 @@
 package br.com.wferreiracosta.louis.controllers;
 
+import br.com.wferreiracosta.louis.exceptions.ValidationError;
 import br.com.wferreiracosta.louis.models.entities.WalletEntity;
-import br.com.wferreiracosta.louis.models.exceptions.ValidationError;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.media.Content;
 import io.swagger.v3.oas.annotations.media.Schema;
