+++ src/main/java/br/com/wferreiracosta/louis/exceptions/StandardError.java
@@ -1,13 +1,10 @@
-package br.com.wferreiracosta.louis.models.exceptions;
+package br.com.wferreiracosta.louis.exceptions;
 
 import io.swagger.v3.oas.annotations.media.Schema;
-import lombok.Builder;
 
 import java.sql.Timestamp;
-import java.util.List;
 
-@Builder
-public record ValidationError(
+public record StandardError(
         @Schema(description = "Error status", example = "404")
         Integer status,
 
@@ -15,10 +12,6 @@ public record ValidationError(
         String message,
 
         @Schema(description = "Error timestamp", example = "2021-03-24 16:48:05.591")
-        Timestamp timestamp,
-
-        @Schema(description = "Errors list")
-        List<FieldMessage> errors
-
+        Timestamp timestamp
 ) {
 }
