+++ src/main/java/br/com/wferreiracosta/louis/exceptions/ValidationError.java
@@ -1,12 +1,14 @@
-package br.com.wferreiracosta.louis.models.exceptions;
+package br.com.wferreiracosta.louis.exceptions;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.Builder;
 
 import java.sql.Timestamp;
+import java.util.List;
 
 @Builder
-public record StandardError(
+public record ValidationError(
+
         @Schema(description = "Error status", example = "404")
         Integer status,
 
@@ -14,6 +16,10 @@ public record StandardError(
         String message,
 
         @Schema(description = "Error timestamp", example = "2021-03-24 16:48:05.591")
-        Timestamp timestamp
+        Timestamp timestamp,
+
+        @Schema(description = "Errors list")
+        List<FieldMessage> errors
+
 ) {
 }
